+++ src/metagen/metaheuristics/genetic/steady_state_genetic_algorithm.py
@@ -20,7 +20,7 @@ from copy import deepcopy
 from typing import Any, Optional, List, Tuple, cast
 
 from metagen.framework import Domain, RelativeAlteration, Solution
-from .ga_tools import (GASolution, yield_two_children, require_crossover,
+from .genetic_tools import (GASolution, yield_two_children, require_crossover,
                        tournament_selection)
 from metagen.metaheuristics.base import Metaheuristic
 from metagen.metaheuristics.tools import random_exploration